pyproject.toml
[project]
name = "nvc-chat-talk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymupdf>=1.27.1",
    "google-api-python-client>=2.100.0",
    "python-dotenv>=1.0.0",
    "langgraph>=0.4",
    "langchain-google-genai>=4.0",
    "langchain-core>=0.3",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "youtube-transcript-api>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]
